// src/types/types.ts
export type Character = {
  id: number;
  name: string;
  status: "Alive" | "Dead" | "unknown";
  species: string;
  type: string;
  gender: "Female" | "Male" | "Genderless" | "unknown";
  origin: {
    name: string;
    url: string;
  };
  location: {
    name: string;
    url: string;
  };
  image: string;
  episode: string[];
  url: string;
  created: string;
};







export type FormsProps = {
  status: string;
  gender: string;
  name: string;
};

export type PaginateProps = {
  countPages: number;
}